+++ app/frontend/src/scss/content/myprofile/ManageFriendContent.scss
