+++ app/frontend/src/scss/content/myprofile/MyProfileContent.scss
@@ -1,4 +1,4 @@
-@import "../../colors.scss";
+@import "../../../colors.scss";
 @import url("https://cdn.rawgit.com/innks/NanumSquareRound/master/nanumsquareround.min.css");
 
 #profile {
